requirements.txt
streamlit
yfinance
pyyaml
pandas
plotly
